Retry cypress/client tests on CI
While it would be nice to be able to resolve all test flakiness, it's proving very difficult. This change adds a single retry to the cypress and client tests (which have been the most flaky). That way, a test suite has to fail twice in a row to register as a failure.